Working on a verify password function
