+++ wc-tracker/src/components/register/actuallogin.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Register() {
+    return(
+        <div>
+            <form onSubmit="submit">
+                <input></input>
+                <button>Submit</button>
+            </form>
+        </div>
+    )
+}
+
+export default Register;
